moved fetching func to correct table view controller
